Move getNoSqlRbacToken to separate file.
